requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==39.0.2
Django==4.1.7
djangorestframework==3.14.0
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
mypy==1.0.1
mypy-extensions==1.0.0
packaging==23.0
pluggy==1.0.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pytest==7.2.2
python-creditcard==0.0.1 #pip install git+https://github.com/maistodos/python-creditcard.git@main
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
xlwt==1.3.0